requirements.txt
apache-airflow
dvc[gs]
flask
scikit-learn
scipy
pandas
numpy
requests